requirements.txt
networkx
matplotlib
google-generativeai
python-dotenv
python-louvain
